Breakup large OcsBuild definition into multiple traits
