+++ modules/edu.gemini.seqexec.web/edu.gemini.seqexec.web.server/src/main/scala/edu/gemini/seqexec/web/server/common/LogInitialization.scala
@@ -51,7 +51,7 @@ trait LogInitialization {
 
   private def makeLogDir(logDir: File): LogConf[Unit] = EitherT {
     IO {
-      (!(logDir.exists() || logDir.mkdirs())).fold(\/-(()), -\/(LogDirNotFound(logDir)))
+      (logDir.exists() || logDir.mkdirs()).fold(\/-(()), -\/(LogDirNotFound(logDir)))
     }
   }
